requirements.txt
fastapi
uvicorn[standard]
python-multipart
qdrant-client
sentence-transformers
groq
ffmpeg-python
imageio-ffmpeg
Pillow
python-dotenv
yt-dlp
